+++ src/components/sessions/StudentSessions.js
@@ -218,7 +218,7 @@ const StudentSessions = ({ open, onClose, student }) => {
                   <MenuItem value="unassigned">
                     <em>Unassigned</em>
                   </MenuItem>
-                  {classes.map((cls) => (
+                  {classes.filter(cls => !cls.disabled).map((cls) => (
                     <MenuItem key={cls.id} value={cls.id}>
                       {cls.name}
                     </MenuItem>